Rollback Avastar to v2.81.23
And adjust snapping at neck for headless body, which has weights at the neck seam to match the mNeck weights in the only replica of the inworld mesh we currently have. The neck seam breaks between system head and mesh body with the bodyfat slider - I can't fix it until I can figure out how the inworld head is rigged for the bodyfat slider.